feat: Convertido para projeto maven
