--- .vscode/tasks.json
@@ -1,19 +1,0 @@
-{
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "label": "Generate ANTLR and Compile Java",
-      "type": "shell",
-      "command": "cmd.exe",
-      "args": [
-        "/c",
-        "java -Xmx500m -cp lib/antlr-4.13.1-complete.jar org.antlr.v4.Tool -Dlanguage=Java -visitor -o src/com/antlrjavacompiler/antlr src/com/antlrjavacompiler/grammar/Grammar.g4 && javac -d output -cp lib/antlr-4.13.1-complete.jar src/com/antlrjavacompiler/antlr/*.java src/com/antlrjavacompiler/Main.java src/com/antlrjavacompiler/exception/*.java src/com/antlrjavacompiler/interpreter/*.java src/com/antlrjavacompiler/classcheck/*java src/com/antlrjavacompiler/tokengenerator/*.java src/com/antlrjavacompiler/ast/dot/*.java src/com/antlrjavacompiler/ast/image/*.java src/com/antlrjavacompiler/ast/viewer/*.java"
-      ],
-      "group": {
-        "kind": "build",
-        "isDefault": true
-      },
-      "problemMatcher": []
-    }
-  ]
-}
