--- src/com/antlrjavacompiler/grammar/.antlr/GrammarParser.java
@@ -1,1237 +1,0 @@
-// Generated from c:/Users/diogo/Desktop/Aplicativos/Programação/Compiladores/compilador_java/src/grammar/Grammar.g4 by ANTLR 4.13.1
-
-package antlr;
-
-import org.antlr.v4.runtime.atn.*;
-import org.antlr.v4.runtime.dfa.DFA;
-import org.antlr.v4.runtime.*;
-import org.antlr.v4.runtime.misc.*;
-import org.antlr.v4.runtime.tree.*;
-import java.util.List;
-import java.util.Iterator;
-import java.util.ArrayList;
-
-@SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast", "CheckReturnValue"})
-public class GrammarParser extends Parser {
-	static { RuntimeMetaData.checkVersion("4.13.1", RuntimeMetaData.VERSION); }
-
-	protected static final DFA[] _decisionToDFA;
-	protected static final PredictionContextCache _sharedContextCache =
-		new PredictionContextCache();
-	public static final int
-		T__0=1, T__1=2, T__2=3, T__3=4, T__4=5, T__5=6, T__6=7, T__7=8, T__8=9, 
-		T__9=10, T__10=11, T__11=12, T__12=13, T__13=14, T__14=15, T__15=16, T__16=17, 
-		T__17=18, T__18=19, T__19=20, T__20=21, T__21=22, T__22=23, T__23=24, 
-		T__24=25, T__25=26, T__26=27, T__27=28, ID=29, INT=30, STRING=31, WS=32, 
-		COMMENT=33;
-	public static final int
-		RULE_program = 0, RULE_declarations = 1, RULE_declaration = 2, RULE_statements = 3, 
-		RULE_statement = 4, RULE_assignment = 5, RULE_read = 6, RULE_write = 7, 
-		RULE_ifStatement = 8, RULE_whileStatement = 9, RULE_block = 10, RULE_expression = 11, 
-		RULE_concatenation = 12, RULE_additiveExpression = 13, RULE_term = 14, 
-		RULE_factor = 15, RULE_logicalExpression = 16, RULE_logicalFactor = 17, 
-		RULE_comparison = 18;
-	private static String[] makeRuleNames() {
-		return new String[] {
-			"program", "declarations", "declaration", "statements", "statement", 
-			"assignment", "read", "write", "ifStatement", "whileStatement", "block", 
-			"expression", "concatenation", "additiveExpression", "term", "factor", 
-			"logicalExpression", "logicalFactor", "comparison"
-		};
-	}
-	public static final String[] ruleNames = makeRuleNames();
-
-	private static String[] makeLiteralNames() {
-		return new String[] {
-			null, "'public'", "'class'", "'{'", "'}'", "'int'", "'string'", "';'", 
-			"'='", "'scanf'", "'('", "')'", "'println'", "'print'", "'if'", "'else'", 
-			"'while'", "'+'", "'-'", "'*'", "'/'", "'&&'", "'||'", "'!'", "'=='", 
-			"'>'", "'<'", "'>='", "'<='"
-		};
-	}
-	private static final String[] _LITERAL_NAMES = makeLiteralNames();
-	private static String[] makeSymbolicNames() {
-		return new String[] {
-			null, null, null, null, null, null, null, null, null, null, null, null, 
-			null, null, null, null, null, null, null, null, null, null, null, null, 
-			null, null, null, null, null, "ID", "INT", "STRING", "WS", "COMMENT"
-		};
-	}
-	private static final String[] _SYMBOLIC_NAMES = makeSymbolicNames();
-	public static final Vocabulary VOCABULARY = new VocabularyImpl(_LITERAL_NAMES, _SYMBOLIC_NAMES);
-
-	/**
-	 * @deprecated Use {@link #VOCABULARY} instead.
-	 */
-	@Deprecated
-	public static final String[] tokenNames;
-	static {
-		tokenNames = new String[_SYMBOLIC_NAMES.length];
-		for (int i = 0; i < tokenNames.length; i++) {
-			tokenNames[i] = VOCABULARY.getLiteralName(i);
-			if (tokenNames[i] == null) {
-				tokenNames[i] = VOCABULARY.getSymbolicName(i);
-			}
-
-			if (tokenNames[i] == null) {
-				tokenNames[i] = "<INVALID>";
-			}
-		}
-	}
-
-	@Override
-	@Deprecated
-	public String[] getTokenNames() {
-		return tokenNames;
-	}
-
-	@Override
-
-	public Vocabulary getVocabulary() {
-		return VOCABULARY;
-	}
-
-	@Override
-	public String getGrammarFileName() { return "Grammar.g4"; }
-
-	@Override
-	public String[] getRuleNames() { return ruleNames; }
-
-	@Override
-	public String getSerializedATN() { return _serializedATN; }
-
-	@Override
-	public ATN getATN() { return _ATN; }
-
-	public GrammarParser(TokenStream input) {
-		super(input);
-		_interp = new ParserATNSimulator(this,_ATN,_decisionToDFA,_sharedContextCache);
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class ProgramContext extends ParserRuleContext {
-		public Token className;
-		public DeclarationsContext declarations() {
-			return getRuleContext(DeclarationsContext.class,0);
-		}
-		public StatementsContext statements() {
-			return getRuleContext(StatementsContext.class,0);
-		}
-		public TerminalNode EOF() { return getToken(GrammarParser.EOF, 0); }
-		public TerminalNode ID() { return getToken(GrammarParser.ID, 0); }
-		public ProgramContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_program; }
-	}
-
-	public final ProgramContext program() throws RecognitionException {
-		ProgramContext _localctx = new ProgramContext(_ctx, getState());
-		enterRule(_localctx, 0, RULE_program);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(38);
-			match(T__0);
-			setState(39);
-			match(T__1);
-			setState(40);
-			((ProgramContext)_localctx).className = match(ID);
-			setState(41);
-			match(T__2);
-			setState(42);
-			declarations();
-			setState(43);
-			statements();
-			setState(44);
-			match(T__3);
-			setState(45);
-			match(EOF);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class DeclarationsContext extends ParserRuleContext {
-		public List<DeclarationContext> declaration() {
-			return getRuleContexts(DeclarationContext.class);
-		}
-		public DeclarationContext declaration(int i) {
-			return getRuleContext(DeclarationContext.class,i);
-		}
-		public DeclarationsContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_declarations; }
-	}
-
-	public final DeclarationsContext declarations() throws RecognitionException {
-		DeclarationsContext _localctx = new DeclarationsContext(_ctx, getState());
-		enterRule(_localctx, 2, RULE_declarations);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(50);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while (_la==T__4 || _la==T__5) {
-				{
-				{
-				setState(47);
-				declaration();
-				}
-				}
-				setState(52);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class DeclarationContext extends ParserRuleContext {
-		public TerminalNode ID() { return getToken(GrammarParser.ID, 0); }
-		public DeclarationContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_declaration; }
-	}
-
-	public final DeclarationContext declaration() throws RecognitionException {
-		DeclarationContext _localctx = new DeclarationContext(_ctx, getState());
-		enterRule(_localctx, 4, RULE_declaration);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(53);
-			_la = _input.LA(1);
-			if ( !(_la==T__4 || _la==T__5) ) {
-			_errHandler.recoverInline(this);
-			}
-			else {
-				if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
-				_errHandler.reportMatch(this);
-				consume();
-			}
-			setState(54);
-			match(ID);
-			setState(55);
-			match(T__6);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class StatementsContext extends ParserRuleContext {
-		public List<StatementContext> statement() {
-			return getRuleContexts(StatementContext.class);
-		}
-		public StatementContext statement(int i) {
-			return getRuleContext(StatementContext.class,i);
-		}
-		public StatementsContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_statements; }
-	}
-
-	public final StatementsContext statements() throws RecognitionException {
-		StatementsContext _localctx = new StatementsContext(_ctx, getState());
-		enterRule(_localctx, 6, RULE_statements);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(60);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while ((((_la) & ~0x3f) == 0 && ((1L << _la) & 536965632L) != 0)) {
-				{
-				{
-				setState(57);
-				statement();
-				}
-				}
-				setState(62);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class StatementContext extends ParserRuleContext {
-		public AssignmentContext assignment() {
-			return getRuleContext(AssignmentContext.class,0);
-		}
-		public ReadContext read() {
-			return getRuleContext(ReadContext.class,0);
-		}
-		public WriteContext write() {
-			return getRuleContext(WriteContext.class,0);
-		}
-		public IfStatementContext ifStatement() {
-			return getRuleContext(IfStatementContext.class,0);
-		}
-		public WhileStatementContext whileStatement() {
-			return getRuleContext(WhileStatementContext.class,0);
-		}
-		public StatementContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_statement; }
-	}
-
-	public final StatementContext statement() throws RecognitionException {
-		StatementContext _localctx = new StatementContext(_ctx, getState());
-		enterRule(_localctx, 8, RULE_statement);
-		try {
-			setState(68);
-			_errHandler.sync(this);
-			switch (_input.LA(1)) {
-			case ID:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(63);
-				assignment();
-				}
-				break;
-			case T__8:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(64);
-				read();
-				}
-				break;
-			case T__11:
-			case T__12:
-				enterOuterAlt(_localctx, 3);
-				{
-				setState(65);
-				write();
-				}
-				break;
-			case T__13:
-				enterOuterAlt(_localctx, 4);
-				{
-				setState(66);
-				ifStatement();
-				}
-				break;
-			case T__15:
-				enterOuterAlt(_localctx, 5);
-				{
-				setState(67);
-				whileStatement();
-				}
-				break;
-			default:
-				throw new NoViableAltException(this);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class AssignmentContext extends ParserRuleContext {
-		public TerminalNode ID() { return getToken(GrammarParser.ID, 0); }
-		public ExpressionContext expression() {
-			return getRuleContext(ExpressionContext.class,0);
-		}
-		public AssignmentContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_assignment; }
-	}
-
-	public final AssignmentContext assignment() throws RecognitionException {
-		AssignmentContext _localctx = new AssignmentContext(_ctx, getState());
-		enterRule(_localctx, 10, RULE_assignment);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(70);
-			match(ID);
-			setState(71);
-			match(T__7);
-			setState(72);
-			expression();
-			setState(73);
-			match(T__6);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class ReadContext extends ParserRuleContext {
-		public TerminalNode ID() { return getToken(GrammarParser.ID, 0); }
-		public ReadContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_read; }
-	}
-
-	public final ReadContext read() throws RecognitionException {
-		ReadContext _localctx = new ReadContext(_ctx, getState());
-		enterRule(_localctx, 12, RULE_read);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(75);
-			match(T__8);
-			setState(76);
-			match(T__9);
-			setState(77);
-			match(ID);
-			setState(78);
-			match(T__10);
-			setState(79);
-			match(T__6);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class WriteContext extends ParserRuleContext {
-		public ExpressionContext expression() {
-			return getRuleContext(ExpressionContext.class,0);
-		}
-		public WriteContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_write; }
-	}
-
-	public final WriteContext write() throws RecognitionException {
-		WriteContext _localctx = new WriteContext(_ctx, getState());
-		enterRule(_localctx, 14, RULE_write);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(81);
-			_la = _input.LA(1);
-			if ( !(_la==T__11 || _la==T__12) ) {
-			_errHandler.recoverInline(this);
-			}
-			else {
-				if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
-				_errHandler.reportMatch(this);
-				consume();
-			}
-			setState(82);
-			match(T__9);
-			setState(83);
-			expression();
-			setState(84);
-			match(T__10);
-			setState(85);
-			match(T__6);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class IfStatementContext extends ParserRuleContext {
-		public LogicalExpressionContext logicalExpression() {
-			return getRuleContext(LogicalExpressionContext.class,0);
-		}
-		public List<BlockContext> block() {
-			return getRuleContexts(BlockContext.class);
-		}
-		public BlockContext block(int i) {
-			return getRuleContext(BlockContext.class,i);
-		}
-		public IfStatementContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_ifStatement; }
-	}
-
-	public final IfStatementContext ifStatement() throws RecognitionException {
-		IfStatementContext _localctx = new IfStatementContext(_ctx, getState());
-		enterRule(_localctx, 16, RULE_ifStatement);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(87);
-			match(T__13);
-			setState(88);
-			match(T__9);
-			setState(89);
-			logicalExpression();
-			setState(90);
-			match(T__10);
-			setState(91);
-			match(T__2);
-			setState(92);
-			block();
-			setState(93);
-			match(T__3);
-			setState(99);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			if (_la==T__14) {
-				{
-				setState(94);
-				match(T__14);
-				setState(95);
-				match(T__2);
-				setState(96);
-				block();
-				setState(97);
-				match(T__3);
-				}
-			}
-
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class WhileStatementContext extends ParserRuleContext {
-		public LogicalExpressionContext logicalExpression() {
-			return getRuleContext(LogicalExpressionContext.class,0);
-		}
-		public BlockContext block() {
-			return getRuleContext(BlockContext.class,0);
-		}
-		public WhileStatementContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_whileStatement; }
-	}
-
-	public final WhileStatementContext whileStatement() throws RecognitionException {
-		WhileStatementContext _localctx = new WhileStatementContext(_ctx, getState());
-		enterRule(_localctx, 18, RULE_whileStatement);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(101);
-			match(T__15);
-			setState(102);
-			match(T__9);
-			setState(103);
-			logicalExpression();
-			setState(104);
-			match(T__10);
-			setState(105);
-			match(T__2);
-			setState(106);
-			block();
-			setState(107);
-			match(T__3);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class BlockContext extends ParserRuleContext {
-		public List<StatementContext> statement() {
-			return getRuleContexts(StatementContext.class);
-		}
-		public StatementContext statement(int i) {
-			return getRuleContext(StatementContext.class,i);
-		}
-		public BlockContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_block; }
-	}
-
-	public final BlockContext block() throws RecognitionException {
-		BlockContext _localctx = new BlockContext(_ctx, getState());
-		enterRule(_localctx, 20, RULE_block);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(110); 
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			do {
-				{
-				{
-				setState(109);
-				statement();
-				}
-				}
-				setState(112); 
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			} while ( (((_la) & ~0x3f) == 0 && ((1L << _la) & 536965632L) != 0) );
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class ExpressionContext extends ParserRuleContext {
-		public ConcatenationContext concatenation() {
-			return getRuleContext(ConcatenationContext.class,0);
-		}
-		public ExpressionContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_expression; }
-	}
-
-	public final ExpressionContext expression() throws RecognitionException {
-		ExpressionContext _localctx = new ExpressionContext(_ctx, getState());
-		enterRule(_localctx, 22, RULE_expression);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(114);
-			concatenation();
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class ConcatenationContext extends ParserRuleContext {
-		public List<AdditiveExpressionContext> additiveExpression() {
-			return getRuleContexts(AdditiveExpressionContext.class);
-		}
-		public AdditiveExpressionContext additiveExpression(int i) {
-			return getRuleContext(AdditiveExpressionContext.class,i);
-		}
-		public ConcatenationContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_concatenation; }
-	}
-
-	public final ConcatenationContext concatenation() throws RecognitionException {
-		ConcatenationContext _localctx = new ConcatenationContext(_ctx, getState());
-		enterRule(_localctx, 24, RULE_concatenation);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(116);
-			additiveExpression();
-			setState(121);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while (_la==T__16) {
-				{
-				{
-				setState(117);
-				match(T__16);
-				setState(118);
-				additiveExpression();
-				}
-				}
-				setState(123);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class AdditiveExpressionContext extends ParserRuleContext {
-		public List<TermContext> term() {
-			return getRuleContexts(TermContext.class);
-		}
-		public TermContext term(int i) {
-			return getRuleContext(TermContext.class,i);
-		}
-		public TerminalNode STRING() { return getToken(GrammarParser.STRING, 0); }
-		public AdditiveExpressionContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_additiveExpression; }
-	}
-
-	public final AdditiveExpressionContext additiveExpression() throws RecognitionException {
-		AdditiveExpressionContext _localctx = new AdditiveExpressionContext(_ctx, getState());
-		enterRule(_localctx, 26, RULE_additiveExpression);
-		int _la;
-		try {
-			int _alt;
-			setState(133);
-			_errHandler.sync(this);
-			switch (_input.LA(1)) {
-			case T__9:
-			case ID:
-			case INT:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(124);
-				term();
-				setState(129);
-				_errHandler.sync(this);
-				_alt = getInterpreter().adaptivePredict(_input,6,_ctx);
-				while ( _alt!=2 && _alt!=org.antlr.v4.runtime.atn.ATN.INVALID_ALT_NUMBER ) {
-					if ( _alt==1 ) {
-						{
-						{
-						setState(125);
-						_la = _input.LA(1);
-						if ( !(_la==T__16 || _la==T__17) ) {
-						_errHandler.recoverInline(this);
-						}
-						else {
-							if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
-							_errHandler.reportMatch(this);
-							consume();
-						}
-						setState(126);
-						term();
-						}
-						} 
-					}
-					setState(131);
-					_errHandler.sync(this);
-					_alt = getInterpreter().adaptivePredict(_input,6,_ctx);
-				}
-				}
-				break;
-			case STRING:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(132);
-				match(STRING);
-				}
-				break;
-			default:
-				throw new NoViableAltException(this);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class TermContext extends ParserRuleContext {
-		public List<FactorContext> factor() {
-			return getRuleContexts(FactorContext.class);
-		}
-		public FactorContext factor(int i) {
-			return getRuleContext(FactorContext.class,i);
-		}
-		public TermContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_term; }
-	}
-
-	public final TermContext term() throws RecognitionException {
-		TermContext _localctx = new TermContext(_ctx, getState());
-		enterRule(_localctx, 28, RULE_term);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(135);
-			factor();
-			setState(140);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while (_la==T__18 || _la==T__19) {
-				{
-				{
-				setState(136);
-				_la = _input.LA(1);
-				if ( !(_la==T__18 || _la==T__19) ) {
-				_errHandler.recoverInline(this);
-				}
-				else {
-					if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
-					_errHandler.reportMatch(this);
-					consume();
-				}
-				setState(137);
-				factor();
-				}
-				}
-				setState(142);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class FactorContext extends ParserRuleContext {
-		public TerminalNode INT() { return getToken(GrammarParser.INT, 0); }
-		public TerminalNode ID() { return getToken(GrammarParser.ID, 0); }
-		public ExpressionContext expression() {
-			return getRuleContext(ExpressionContext.class,0);
-		}
-		public FactorContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_factor; }
-	}
-
-	public final FactorContext factor() throws RecognitionException {
-		FactorContext _localctx = new FactorContext(_ctx, getState());
-		enterRule(_localctx, 30, RULE_factor);
-		try {
-			setState(149);
-			_errHandler.sync(this);
-			switch (_input.LA(1)) {
-			case INT:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(143);
-				match(INT);
-				}
-				break;
-			case ID:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(144);
-				match(ID);
-				}
-				break;
-			case T__9:
-				enterOuterAlt(_localctx, 3);
-				{
-				setState(145);
-				match(T__9);
-				setState(146);
-				expression();
-				setState(147);
-				match(T__10);
-				}
-				break;
-			default:
-				throw new NoViableAltException(this);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class LogicalExpressionContext extends ParserRuleContext {
-		public List<LogicalFactorContext> logicalFactor() {
-			return getRuleContexts(LogicalFactorContext.class);
-		}
-		public LogicalFactorContext logicalFactor(int i) {
-			return getRuleContext(LogicalFactorContext.class,i);
-		}
-		public LogicalExpressionContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_logicalExpression; }
-	}
-
-	public final LogicalExpressionContext logicalExpression() throws RecognitionException {
-		LogicalExpressionContext _localctx = new LogicalExpressionContext(_ctx, getState());
-		enterRule(_localctx, 32, RULE_logicalExpression);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(151);
-			logicalFactor();
-			setState(156);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while (_la==T__20 || _la==T__21) {
-				{
-				{
-				setState(152);
-				_la = _input.LA(1);
-				if ( !(_la==T__20 || _la==T__21) ) {
-				_errHandler.recoverInline(this);
-				}
-				else {
-					if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
-					_errHandler.reportMatch(this);
-					consume();
-				}
-				setState(153);
-				logicalFactor();
-				}
-				}
-				setState(158);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class LogicalFactorContext extends ParserRuleContext {
-		public ComparisonContext comparison() {
-			return getRuleContext(ComparisonContext.class,0);
-		}
-		public LogicalExpressionContext logicalExpression() {
-			return getRuleContext(LogicalExpressionContext.class,0);
-		}
-		public LogicalFactorContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_logicalFactor; }
-	}
-
-	public final LogicalFactorContext logicalFactor() throws RecognitionException {
-		LogicalFactorContext _localctx = new LogicalFactorContext(_ctx, getState());
-		enterRule(_localctx, 34, RULE_logicalFactor);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(160);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			if (_la==T__22) {
-				{
-				setState(159);
-				match(T__22);
-				}
-			}
-
-			setState(167);
-			_errHandler.sync(this);
-			switch ( getInterpreter().adaptivePredict(_input,12,_ctx) ) {
-			case 1:
-				{
-				setState(162);
-				comparison();
-				}
-				break;
-			case 2:
-				{
-				setState(163);
-				match(T__9);
-				setState(164);
-				logicalExpression();
-				setState(165);
-				match(T__10);
-				}
-				break;
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	@SuppressWarnings("CheckReturnValue")
-	public static class ComparisonContext extends ParserRuleContext {
-		public List<ExpressionContext> expression() {
-			return getRuleContexts(ExpressionContext.class);
-		}
-		public ExpressionContext expression(int i) {
-			return getRuleContext(ExpressionContext.class,i);
-		}
-		public ComparisonContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_comparison; }
-	}
-
-	public final ComparisonContext comparison() throws RecognitionException {
-		ComparisonContext _localctx = new ComparisonContext(_ctx, getState());
-		enterRule(_localctx, 36, RULE_comparison);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(169);
-			expression();
-			setState(170);
-			_la = _input.LA(1);
-			if ( !((((_la) & ~0x3f) == 0 && ((1L << _la) & 520093696L) != 0)) ) {
-			_errHandler.recoverInline(this);
-			}
-			else {
-				if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
-				_errHandler.reportMatch(this);
-				consume();
-			}
-			setState(171);
-			expression();
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static final String _serializedATN =
-		"\u0004\u0001!\u00ae\u0002\u0000\u0007\u0000\u0002\u0001\u0007\u0001\u0002"+
-		"\u0002\u0007\u0002\u0002\u0003\u0007\u0003\u0002\u0004\u0007\u0004\u0002"+
-		"\u0005\u0007\u0005\u0002\u0006\u0007\u0006\u0002\u0007\u0007\u0007\u0002"+
-		"\b\u0007\b\u0002\t\u0007\t\u0002\n\u0007\n\u0002\u000b\u0007\u000b\u0002"+
-		"\f\u0007\f\u0002\r\u0007\r\u0002\u000e\u0007\u000e\u0002\u000f\u0007\u000f"+
-		"\u0002\u0010\u0007\u0010\u0002\u0011\u0007\u0011\u0002\u0012\u0007\u0012"+
-		"\u0001\u0000\u0001\u0000\u0001\u0000\u0001\u0000\u0001\u0000\u0001\u0000"+
-		"\u0001\u0000\u0001\u0000\u0001\u0000\u0001\u0001\u0005\u00011\b\u0001"+
-		"\n\u0001\f\u00014\t\u0001\u0001\u0002\u0001\u0002\u0001\u0002\u0001\u0002"+
-		"\u0001\u0003\u0005\u0003;\b\u0003\n\u0003\f\u0003>\t\u0003\u0001\u0004"+
-		"\u0001\u0004\u0001\u0004\u0001\u0004\u0001\u0004\u0003\u0004E\b\u0004"+
-		"\u0001\u0005\u0001\u0005\u0001\u0005\u0001\u0005\u0001\u0005\u0001\u0006"+
-		"\u0001\u0006\u0001\u0006\u0001\u0006\u0001\u0006\u0001\u0006\u0001\u0007"+
-		"\u0001\u0007\u0001\u0007\u0001\u0007\u0001\u0007\u0001\u0007\u0001\b\u0001"+
-		"\b\u0001\b\u0001\b\u0001\b\u0001\b\u0001\b\u0001\b\u0001\b\u0001\b\u0001"+
-		"\b\u0001\b\u0003\bd\b\b\u0001\t\u0001\t\u0001\t\u0001\t\u0001\t\u0001"+
-		"\t\u0001\t\u0001\t\u0001\n\u0004\no\b\n\u000b\n\f\np\u0001\u000b\u0001"+
-		"\u000b\u0001\f\u0001\f\u0001\f\u0005\fx\b\f\n\f\f\f{\t\f\u0001\r\u0001"+
-		"\r\u0001\r\u0005\r\u0080\b\r\n\r\f\r\u0083\t\r\u0001\r\u0003\r\u0086\b"+
-		"\r\u0001\u000e\u0001\u000e\u0001\u000e\u0005\u000e\u008b\b\u000e\n\u000e"+
-		"\f\u000e\u008e\t\u000e\u0001\u000f\u0001\u000f\u0001\u000f\u0001\u000f"+
-		"\u0001\u000f\u0001\u000f\u0003\u000f\u0096\b\u000f\u0001\u0010\u0001\u0010"+
-		"\u0001\u0010\u0005\u0010\u009b\b\u0010\n\u0010\f\u0010\u009e\t\u0010\u0001"+
-		"\u0011\u0003\u0011\u00a1\b\u0011\u0001\u0011\u0001\u0011\u0001\u0011\u0001"+
-		"\u0011\u0001\u0011\u0003\u0011\u00a8\b\u0011\u0001\u0012\u0001\u0012\u0001"+
-		"\u0012\u0001\u0012\u0001\u0012\u0000\u0000\u0013\u0000\u0002\u0004\u0006"+
-		"\b\n\f\u000e\u0010\u0012\u0014\u0016\u0018\u001a\u001c\u001e \"$\u0000"+
-		"\u0006\u0001\u0000\u0005\u0006\u0001\u0000\f\r\u0001\u0000\u0011\u0012"+
-		"\u0001\u0000\u0013\u0014\u0001\u0000\u0015\u0016\u0001\u0000\u0018\u001c"+
-		"\u00ab\u0000&\u0001\u0000\u0000\u0000\u00022\u0001\u0000\u0000\u0000\u0004"+
-		"5\u0001\u0000\u0000\u0000\u0006<\u0001\u0000\u0000\u0000\bD\u0001\u0000"+
-		"\u0000\u0000\nF\u0001\u0000\u0000\u0000\fK\u0001\u0000\u0000\u0000\u000e"+
-		"Q\u0001\u0000\u0000\u0000\u0010W\u0001\u0000\u0000\u0000\u0012e\u0001"+
-		"\u0000\u0000\u0000\u0014n\u0001\u0000\u0000\u0000\u0016r\u0001\u0000\u0000"+
-		"\u0000\u0018t\u0001\u0000\u0000\u0000\u001a\u0085\u0001\u0000\u0000\u0000"+
-		"\u001c\u0087\u0001\u0000\u0000\u0000\u001e\u0095\u0001\u0000\u0000\u0000"+
-		" \u0097\u0001\u0000\u0000\u0000\"\u00a0\u0001\u0000\u0000\u0000$\u00a9"+
-		"\u0001\u0000\u0000\u0000&\'\u0005\u0001\u0000\u0000\'(\u0005\u0002\u0000"+
-		"\u0000()\u0005\u001d\u0000\u0000)*\u0005\u0003\u0000\u0000*+\u0003\u0002"+
-		"\u0001\u0000+,\u0003\u0006\u0003\u0000,-\u0005\u0004\u0000\u0000-.\u0005"+
-		"\u0000\u0000\u0001.\u0001\u0001\u0000\u0000\u0000/1\u0003\u0004\u0002"+
-		"\u00000/\u0001\u0000\u0000\u000014\u0001\u0000\u0000\u000020\u0001\u0000"+
-		"\u0000\u000023\u0001\u0000\u0000\u00003\u0003\u0001\u0000\u0000\u0000"+
-		"42\u0001\u0000\u0000\u000056\u0007\u0000\u0000\u000067\u0005\u001d\u0000"+
-		"\u000078\u0005\u0007\u0000\u00008\u0005\u0001\u0000\u0000\u00009;\u0003"+
-		"\b\u0004\u0000:9\u0001\u0000\u0000\u0000;>\u0001\u0000\u0000\u0000<:\u0001"+
-		"\u0000\u0000\u0000<=\u0001\u0000\u0000\u0000=\u0007\u0001\u0000\u0000"+
-		"\u0000><\u0001\u0000\u0000\u0000?E\u0003\n\u0005\u0000@E\u0003\f\u0006"+
-		"\u0000AE\u0003\u000e\u0007\u0000BE\u0003\u0010\b\u0000CE\u0003\u0012\t"+
-		"\u0000D?\u0001\u0000\u0000\u0000D@\u0001\u0000\u0000\u0000DA\u0001\u0000"+
-		"\u0000\u0000DB\u0001\u0000\u0000\u0000DC\u0001\u0000\u0000\u0000E\t\u0001"+
-		"\u0000\u0000\u0000FG\u0005\u001d\u0000\u0000GH\u0005\b\u0000\u0000HI\u0003"+
-		"\u0016\u000b\u0000IJ\u0005\u0007\u0000\u0000J\u000b\u0001\u0000\u0000"+
-		"\u0000KL\u0005\t\u0000\u0000LM\u0005\n\u0000\u0000MN\u0005\u001d\u0000"+
-		"\u0000NO\u0005\u000b\u0000\u0000OP\u0005\u0007\u0000\u0000P\r\u0001\u0000"+
-		"\u0000\u0000QR\u0007\u0001\u0000\u0000RS\u0005\n\u0000\u0000ST\u0003\u0016"+
-		"\u000b\u0000TU\u0005\u000b\u0000\u0000UV\u0005\u0007\u0000\u0000V\u000f"+
-		"\u0001\u0000\u0000\u0000WX\u0005\u000e\u0000\u0000XY\u0005\n\u0000\u0000"+
-		"YZ\u0003 \u0010\u0000Z[\u0005\u000b\u0000\u0000[\\\u0005\u0003\u0000\u0000"+
-		"\\]\u0003\u0014\n\u0000]c\u0005\u0004\u0000\u0000^_\u0005\u000f\u0000"+
-		"\u0000_`\u0005\u0003\u0000\u0000`a\u0003\u0014\n\u0000ab\u0005\u0004\u0000"+
-		"\u0000bd\u0001\u0000\u0000\u0000c^\u0001\u0000\u0000\u0000cd\u0001\u0000"+
-		"\u0000\u0000d\u0011\u0001\u0000\u0000\u0000ef\u0005\u0010\u0000\u0000"+
-		"fg\u0005\n\u0000\u0000gh\u0003 \u0010\u0000hi\u0005\u000b\u0000\u0000"+
-		"ij\u0005\u0003\u0000\u0000jk\u0003\u0014\n\u0000kl\u0005\u0004\u0000\u0000"+
-		"l\u0013\u0001\u0000\u0000\u0000mo\u0003\b\u0004\u0000nm\u0001\u0000\u0000"+
-		"\u0000op\u0001\u0000\u0000\u0000pn\u0001\u0000\u0000\u0000pq\u0001\u0000"+
-		"\u0000\u0000q\u0015\u0001\u0000\u0000\u0000rs\u0003\u0018\f\u0000s\u0017"+
-		"\u0001\u0000\u0000\u0000ty\u0003\u001a\r\u0000uv\u0005\u0011\u0000\u0000"+
-		"vx\u0003\u001a\r\u0000wu\u0001\u0000\u0000\u0000x{\u0001\u0000\u0000\u0000"+
-		"yw\u0001\u0000\u0000\u0000yz\u0001\u0000\u0000\u0000z\u0019\u0001\u0000"+
-		"\u0000\u0000{y\u0001\u0000\u0000\u0000|\u0081\u0003\u001c\u000e\u0000"+
-		"}~\u0007\u0002\u0000\u0000~\u0080\u0003\u001c\u000e\u0000\u007f}\u0001"+
-		"\u0000\u0000\u0000\u0080\u0083\u0001\u0000\u0000\u0000\u0081\u007f\u0001"+
-		"\u0000\u0000\u0000\u0081\u0082\u0001\u0000\u0000\u0000\u0082\u0086\u0001"+
-		"\u0000\u0000\u0000\u0083\u0081\u0001\u0000\u0000\u0000\u0084\u0086\u0005"+
-		"\u001f\u0000\u0000\u0085|\u0001\u0000\u0000\u0000\u0085\u0084\u0001\u0000"+
-		"\u0000\u0000\u0086\u001b\u0001\u0000\u0000\u0000\u0087\u008c\u0003\u001e"+
-		"\u000f\u0000\u0088\u0089\u0007\u0003\u0000\u0000\u0089\u008b\u0003\u001e"+
-		"\u000f\u0000\u008a\u0088\u0001\u0000\u0000\u0000\u008b\u008e\u0001\u0000"+
-		"\u0000\u0000\u008c\u008a\u0001\u0000\u0000\u0000\u008c\u008d\u0001\u0000"+
-		"\u0000\u0000\u008d\u001d\u0001\u0000\u0000\u0000\u008e\u008c\u0001\u0000"+
-		"\u0000\u0000\u008f\u0096\u0005\u001e\u0000\u0000\u0090\u0096\u0005\u001d"+
-		"\u0000\u0000\u0091\u0092\u0005\n\u0000\u0000\u0092\u0093\u0003\u0016\u000b"+
-		"\u0000\u0093\u0094\u0005\u000b\u0000\u0000\u0094\u0096\u0001\u0000\u0000"+
-		"\u0000\u0095\u008f\u0001\u0000\u0000\u0000\u0095\u0090\u0001\u0000\u0000"+
-		"\u0000\u0095\u0091\u0001\u0000\u0000\u0000\u0096\u001f\u0001\u0000\u0000"+
-		"\u0000\u0097\u009c\u0003\"\u0011\u0000\u0098\u0099\u0007\u0004\u0000\u0000"+
-		"\u0099\u009b\u0003\"\u0011\u0000\u009a\u0098\u0001\u0000\u0000\u0000\u009b"+
-		"\u009e\u0001\u0000\u0000\u0000\u009c\u009a\u0001\u0000\u0000\u0000\u009c"+
-		"\u009d\u0001\u0000\u0000\u0000\u009d!\u0001\u0000\u0000\u0000\u009e\u009c"+
-		"\u0001\u0000\u0000\u0000\u009f\u00a1\u0005\u0017\u0000\u0000\u00a0\u009f"+
-		"\u0001\u0000\u0000\u0000\u00a0\u00a1\u0001\u0000\u0000\u0000\u00a1\u00a7"+
-		"\u0001\u0000\u0000\u0000\u00a2\u00a8\u0003$\u0012\u0000\u00a3\u00a4\u0005"+
-		"\n\u0000\u0000\u00a4\u00a5\u0003 \u0010\u0000\u00a5\u00a6\u0005\u000b"+
-		"\u0000\u0000\u00a6\u00a8\u0001\u0000\u0000\u0000\u00a7\u00a2\u0001\u0000"+
-		"\u0000\u0000\u00a7\u00a3\u0001\u0000\u0000\u0000\u00a8#\u0001\u0000\u0000"+
-		"\u0000\u00a9\u00aa\u0003\u0016\u000b\u0000\u00aa\u00ab\u0007\u0005\u0000"+
-		"\u0000\u00ab\u00ac\u0003\u0016\u000b\u0000\u00ac%\u0001\u0000\u0000\u0000"+
-		"\r2<Dcpy\u0081\u0085\u008c\u0095\u009c\u00a0\u00a7";
-	public static final ATN _ATN =
-		new ATNDeserializer().deserialize(_serializedATN.toCharArray());
-	static {
-		_decisionToDFA = new DFA[_ATN.getNumberOfDecisions()];
-		for (int i = 0; i < _ATN.getNumberOfDecisions(); i++) {
-			_decisionToDFA[i] = new DFA(_ATN.getDecisionState(i), i);
-		}
-	}
-}
